Feature/resolve merge conflict v2 (#340)
* Move vespa search tests & the search_fixtures they use under dedicated sub-folder  (#334)

* Move vespa search tests under dedicated vespa folder

* Move /search_fixtures under vespa search folder & rename to fixtures

* Bump to 1.14.20

* Move vespa search result order tests into a separate file (#335)

* Move vespa search tests under dedicated vespa folder

* Move /search_fixtures under vespa search folder & rename to fixtures

* Bump to 1.14.20

* Move vespa search result order tests to separate file

* Bump to 1.14.19

* Move continuation token vespa search tests to separate file (#336)

* Move vespa search tests under dedicated vespa folder

* Move /search_fixtures under vespa search folder & rename to fixtures

* Bump to 1.14.20

* Move vespa search result order tests to separate file

* Bump to 1.14.19

* Move vespa search continuation token tests to separate file

* Group pagination and continuation token tests

* Move keyword and range vespa search tests into separate file (#337)

* Move vespa search tests under dedicated vespa folder

* Move /search_fixtures under vespa search folder & rename to fixtures

* Bump to 1.14.20

* Move vespa search result order tests to separate file

* Bump to 1.14.19

* Move vespa search continuation token tests to separate file

* Move keyword and range vespa search tests into separate file

* Delete test_vespa_search_cont_tokens.py

* Move _make_search_request into vespa search setup

* Move vespa search tests for ignoring special chars & case to separate file (#338)

* Move data download tests into parent folder

* Move query insensitivity & special chars ignoring tests out

* Rename from test_vespasearch

* Bump to 1.14.20

* Removing refactored file.

* Adding back in the changes from the test_vespasearch.

---------
