+++ makefile-docker.defs
@@ -80,17 +80,17 @@ vespa_healthy:
 .ONESHELL:
 vespa_deploy_schema:
 	vespa config set target local
-	@vespa deploy tests/search/search_fixtures/vespa_test_schema --wait 300
+	@vespa deploy tests/search/vespa/fixtures/vespa_test_schema --wait 300
 
 .ONESHELL:
 vespa_load_data:
 	vespa config set target local
-	vespa feed --progress=3 tests/search/search_fixtures/vespa_search_weights.json
-	vespa feed --progress=3 tests/search/search_fixtures/vespa_family_document.json
-	vespa feed --progress=3 tests/search/search_fixtures/vespa_document_passage.json
+	vespa feed --progress=3 tests/search/vespa/fixtures/vespa_search_weights.json
+	vespa feed --progress=3 tests/search/vespa/fixtures/vespa_family_document.json
+	vespa feed --progress=3 tests/search/vespa/fixtures/vespa_document_passage.json
 
 vespa_setup: vespa_confirm_cli_installed vespa_healthy vespa_deploy_schema vespa_load_data
-	# Deploys a vespa application to a local vespa container and loads search_fixtures
+	# Deploys a vespa application to a local vespa container and loads search fixtures
 
 .ONESHELL:
 test_search:
