+++ tests/search/vespa/fixtures/vespa_test_schema/query-profiles/default.xml
