+++ tests/search/vespa/fixtures/vespa_test_schema/schemas/document_passage.sd
