+++ tests/search/vespa/fixtures/vespa_test_schema/schemas/search_weights.sd
