+++ tests/search/vespa/fixtures/vespa_test_schema/services.xml
