+++ tests/search/vespa/test_this_vespa_search_download.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 import pytest
 
 from app.api.api_v1.routers import search
-from tests.search.setup_search_tests import _populate_db_families
+from tests.search.vespa.setup_search_tests import _populate_db_families
 
 SEARCH_ENDPOINT = "/api/v1/searches"
 CSV_DOWNLOAD_ENDPOINT = "/api/v1/searches/download-csv"
