+++ tests/search/vespa/test_whole_database_download.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
-from tests.search.setup_search_tests import _populate_db_families
+from tests.search.vespa.setup_search_tests import _populate_db_families
 
 ALL_DATA_DOWNLOAD_ENDPOINT = "/api/v1/searches/download-all-data"
